Added Tree Drag n Drop
Elements in the tree can now be drag and dropped above other elements in the tree to change the order. Also works with elements that have child items

Dragging a tree element directly onto another element in the tree, makes it become a child of that element.

Also added visual feedback for both drag events. dragging above, shows a line visually either above or below the targeted element and dragging onto an element, changes the background of the target element.

Not that this works instantly and is currently not recorded in the undo history.