Added Prefix and extension overwrite for Textures
On the Options Panel, there are 2 new input fields. One prepents the input text, when selecting a texture for a frame (useful for automatically setting the correct folder structure that holds the UI Elements in your warcraft 3 mod. For example my UI Elements are all in the UI folder inside the root wc3 mpq. Because of that, whenever I select a texture for a frame, I need to manually append UI/ to it, which now can be done automatically.

The other input fields overwrites the file extension. My UI Files are all tga inside the map, but I use .png files for creating the layout. With this set, you can automatically change the .png for the Texture (WE) input into whatever you want.