Added new Input fields to set callbacks for buttons
There are 3 new Input fields for MouseClick,MouseEnter and MouseLeave for Button Type Frames (also for SimpleButtons). With these, you can define Function names, that will automatically be set as the callback for the respective event on that frame. When the fields are left empty, the respective event is not registered / created as export code.

Also texture input is removed / disabled for SimpleButtons, since they are invisible by default and cant have a texture (need to work with an additional backdrop and overlay the SimpleButton)